+++ Assets/Resources/Effects.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c54d29105d754e119f98b0affe1608a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
