+++ Assets/Resources/Effects/GunEffect.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1f7066b33da6f4e7887fdb2215c140fb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
