--- Assets/Standard Assets/Characters/FirstPersonCharacter/Scripts/ShotController.cs
@@ -1,21 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ShotController : MonoBehaviour {
-
-	private Ray shot;
-	private RaycastHit shotHit;
-
-	// Use this for initialization
-	void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-		if(Input.GetMouseButtonDown(0)){
-			shot = 
-		}
-	}
-}
